requirements.txt
streamlit==1.28.0
langchain-core>=0.3.0
langchain-google-genai>=2.0.0
langchain-community>=0.3.0
langsmith>=0.1.75
transformers==4.42.0
torch==2.5.1
sentence-transformers==2.6.1
faiss-cpu==1.12.0
pypdf==3.17.0
python-dotenv==1.0.0
tiktoken>=0.5.2
